refactor: domain, controller 폴더 생성
